+++ webapp/channels/src/actions/notification_actions.test.js
@@ -5,8 +5,8 @@ import testConfigureStore from 'tests/test_store';
 
 import {getHistory} from 'utils/browser_history';
 import Constants, {NotificationLevels, UserStatuses} from 'utils/constants';
+import * as NotificationSounds from 'utils/notification_sounds';
 import * as utils from 'utils/notifications';
-import * as baseUtils from 'utils/utils';
 
 import {sendDesktopNotification} from './notification_actions';
 
@@ -22,7 +22,7 @@ describe('notification_actions', () => {
 
         beforeEach(() => {
             spy = jest.spyOn(utils, 'showNotification');
-            baseUtils.ding = jest.fn();
+            NotificationSounds.ding = jest.fn();
 
             crt = {
                 user_id: 'current_user_id',
@@ -315,7 +315,7 @@ describe('notification_actions', () => {
         });
 
         test('should default sound when no sound is specified', () => {
-            const dingSpy = jest.spyOn(baseUtils, 'ding');
+            const dingSpy = jest.spyOn(NotificationSounds, 'ding');
             baseState.entities.users.profiles.current_user_id.notify_props.desktop_sound = 'true';
             const store = testConfigureStore(baseState);
             return store.dispatch(sendDesktopNotification(post, msgProps)).then(() => {
@@ -324,7 +324,7 @@ describe('notification_actions', () => {
         });
 
         test('should use specified sound when specified', () => {
-            const dingSpy = jest.spyOn(baseUtils, 'ding');
+            const dingSpy = jest.spyOn(NotificationSounds, 'ding');
             baseState.entities.users.profiles.current_user_id.notify_props.desktop_sound = 'true';
             baseState.entities.users.profiles.current_user_id.notify_props.desktop_notification_sound = 'Crackle';
             const store = testConfigureStore(baseState);
